Refactor CalculationAuthorizedRequest into Templates\CalculationAuthorizedRequest
